HW5: refactoring tests
